+++ src/core/db/utxo_db.h
@@ -1,7 +1,18 @@
 #ifndef FCOIN_UTXO_DB_H
 #define FCOIN_UTXO_DB_H
 
-class utxo_db {
+#include "db.h"
+#include "../transaction.h"
+
+class utxo_db : public db<transaction_output>{
+public:
+    utxo_db() : db(UTXO_DB){}
+    ~utxo_db() = default;
+
+    void init() override;
+    void clear() override;
+    void put(const transaction_output &obj) override;
+    transaction_output pull(const str &hash) override;
 
 };